requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
requests==2.31.0
python-dotenv==1.0.1

# Google / Gemini
google-generativeai==0.7.2
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client==2.125.0

# (Optional) Personal KB / RAG features
chromadb
beautifulsoup4
pypdf

slack-sdk==3.27.0
